+++ src/picotorrent/ui/mainframe.hpp
@@ -77,6 +77,7 @@ namespace Models
         void OnFileAddTorrent(wxCommandEvent&);
         void OnFileCreateTorrent(wxCommandEvent&);
         void OnFileImportQBittorrent(wxCommandEvent&);
+        void OnFileImportPicoTorrent(wxCommandEvent&);
         void OnHelpAbout(wxCommandEvent&);
         void OnViewHelp(wxCommandEvent&);
         void OnIconize(wxIconizeEvent&);